Fix a very obscure bug that popped up in some very restricted circumstances.  copy_all_to_FFSBuffer() is designed to take an FFSEncodeVector and an FFSBuffer.  The vector might include pointers both into the FFSBuffer and into external application memory.  The Buffer should be extended and all application data should be copied into it.  Obviously the buffer memory may be realloc()'s in this process, so during processing the pointers that already point into the buffer are converted into offsets.  Then the external data is added with add_to_tmp_buffer(), and when all is done we convert back from offsets to real addresses.  At an in-between stage, the offsets were kept in the iov_base structure, as were the external addresses) and determined which were pointers and which were offsets by seeing which ones were smaller than the buffer size.  This would succeed as long as the external addresses were always larger than the buffer size, which seemed guaranteed, but happened not to always be the case on the 32-bit Nitrogen6x.  SO, we now track explicitly which ones are offsets and which ones are addresses in a separate array.  Sigh.  Also start using long as offsets rather than int, just 'cause we've been trying to support really huge FFS blocks, regardless of their likely use.
git-svn-id: http://svn.research.cc.gatech.edu/kaos/ffs/trunk@20980 8ba6a8b9-845b-0410-94f0-974be2f5df47

+++ ffs/ffs_internal.h
@@ -29,8 +29,8 @@ extern int bswap_32(int l);
 
 struct _FFSBuffer {
     void *tmp_buffer;
-    int tmp_buffer_size;
-    int tmp_buffer_in_use_size;
+    long tmp_buffer_size;
+    long tmp_buffer_in_use_size;
 };
 
 typedef struct _internal_iovec {
